+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Release changelog are available here : https://gitlab.inria.fr/croco-ocean/croco
 
 ### Added
 
+- Issue #250 : Add dredging feature in MUSTANG
+
 ### Fixed
 
 - Issue #239 : back to previous default option in create_config
@@ -35,6 +37,8 @@ Release changelog are available here : https://gitlab.inria.fr/croco-ocean/croco
 - Issue #281 : optimization of the PISCES code on
   - representation of the lability of the particle pool
   - several optimizations to the calculation of certain variables (performance).
+
+- Issue #250 : for MUSTANG initialisation, review of namelist parameters
 
 - Issue #163 : for MUSTANG output, 
   - avoid possibility of overlapping in vname by